chore(components): restructure
