+++ src/components/LabelInfo/LabelInfo.module.scss
@@ -1,4 +1,4 @@
-@import '../../../stylesheets/base';
+@import '../../stylesheets/base';
 
 .label-info {
   @apply border border-solid rounded;
