+++ src/components/LabelList/LabelInfoList/LabelInfoList.tsx
@@ -2,7 +2,7 @@ import classNames from 'classnames';
 import React from 'react';
 import type { TLabelsGroupItem } from '../../../helpers/labelList';
 import { IRemoteSubscription } from '../../../models/Subscription';
-import LabelInfo from '../LabelInfo/LabelInfo';
+import LabelInfo from '../../LabelInfo/LabelInfo';
 import styles from './LabelInfoList.module.scss';
 
 interface IProps {
