+++ src/components/RemoveLabelButton/RemoveLabelButton.tsx
@@ -1,15 +1,15 @@
 import { render } from 'mustache';
 import React, { useCallback } from 'react';
-import cache from '../../../../cache';
-import { EventAction, EventCategory } from '../../../../constants/ga';
-import * as TEXTS from '../../../../constants/texts';
-import * as gtag from '../../../../helpers/gtag';
-import type { ILabel } from '../../../../models/Label';
-import { actions as personalActions } from '../../../../store/slices/personal';
-import { useTypedDispatch } from '../../../../store/store';
-import * as questions from '../../../../templates/questions';
-import { IconName } from '../../../Icon/types';
-import IconButton from '../../../IconButton/IconButton';
+import cache from '../../cache';
+import { EventAction, EventCategory } from '../../constants/ga';
+import * as TEXTS from '../../constants/texts';
+import * as gtag from '../../helpers/gtag';
+import type { ILabel } from '../../models/Label';
+import { actions as personalActions } from '../../store/slices/personal';
+import { useTypedDispatch } from '../../store/store';
+import * as questions from '../../templates/questions';
+import { IconName } from '../Icon/types';
+import IconButton from '../IconButton/IconButton';
 
 interface IProps {
   user: string;
